requirements.txt
asgiref==3.5.2
certifi==2022.6.15
distlib==0.3.5
Django==4.0.6
djangorestframework==3.14.0
filelock==3.4.2
platformdirs==2.5.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.15.1
virtualenv-clone==0.5.7
